Changed css measurement units (almost all to rem).
Changed all css measurement units to rem, with the exception of media
queries (in px), and the html root font-size (in px).

The later is set to be 1px because of yet unexplained behavior from
chrome v56.0.2924.87 currently tested with while fine when tested on
the rest (firefox v51.0.1, Edge v38.14393.0.0, IE v11.576.14393.0).

As an aside, pruned some code here and there too.