Set CDN-ress dependencies to auto download with npm.
* Switched most assets to npm packages / github repositories.
* Set up a preinstall script for lightweight dependencies download.
Works by specifying appropriate destination path inside package.json
'pm-' prefixed attribute.

- Preinstall script notes:
Use of the prescript is an alternative to npm downloading a full
package, with several useless files for your use case, just so you can
use one or two of them.
This works most well with assets available on a CDN, with no relative
dependencies (exp. bootstrap@3 requires fonts in a relative directory).

Currently used for jquery, because only a single file is required.
Also used for d3.min.js, because it's a single file to be required, and
because npm d3 package failed to load in the first place, because of
one of its dependencies' (useless in this use case) bug.